requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.3.0
python-dotenv>=1.0.0
jsonlines>=3.1.0
plotly>=5.15.0
pathlib2>=2.3.7
urllib3>=1.26.0